Removed web + cleanup

+++ VsaApp/app/src/main/java/de/lohl1kohl/vsaapp/jobs/EndJob.java
@@ -6,9 +6,7 @@ import android.content.SharedPreferences;
 import android.media.AudioManager;
 import android.preference.PreferenceManager;
 import android.support.annotation.NonNull;
-
 import com.evernote.android.job.Job;
-
 import de.lohl1kohl.vsaapp.LoadingActivity;
 import de.lohl1kohl.vsaapp.holders.SpHolder;
 
@@ -23,7 +21,7 @@ public class EndJob extends Job {
         final AudioManager mode = (AudioManager) getContext().getSystemService(Context.AUDIO_SERVICE);
         mode.setRingerMode(sharedPreferences.getInt("ringer_mode", AudioManager.RINGER_MODE_VIBRATE));
         int maxVolume = mode.getStreamMaxVolume(AudioManager.STREAM_MUSIC);
-        mode.setStreamVolume(AudioManager.STREAM_MUSIC, sharedPreferences.getInt("media_mode", (int) maxVolume / 2), 0);
+        mode.setStreamVolume(AudioManager.STREAM_MUSIC, sharedPreferences.getInt("media_mode", maxVolume / 2), 0);
 
         NotificationManager notificationManager = (NotificationManager) getContext().getSystemService(Context.NOTIFICATION_SERVICE);
         notificationManager.cancel(-1);
